+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - fixed access of `output_sampledata` flag of `wai.spectralio.adams.Writer`
 - added setter for sample data in `Spectrum` class
 - the NIR Reader no longer skips the reference values
+- serialization of constituents in FOSS NIR header is now more robust
 
 
 0.0.1 (2025-04-14)